Refactor: CurrencyItem: move inside UI folder
