+++ src/components/Currency/UI/CurrencyItem.tsx
